+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- fix: Failing MacOS build #530
 - ref: Session values are unsigned #527
 
 ## 5.0.4