+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - fix: Change SentryScope setTagValue to NSString #524
+- fix: Failing MacOS build #530
 - ref: Session values are unsigned #527
 
 ## 5.0.4